RFC-0037: integers have a width, and a literal takes the width its use demands
i8..i64 and u8..u64 replace Int and Byte. TyTerm::Int(IntTy); an integer
literal is a variable bounded to integer widths, i64 where nothing
narrows it, signed-only when negated, and checked against its value once
the width is known. Arithmetic runs at the operands' width with checked
overflow; a space lays an integer out in its width's bytes; JSON reads a
number as i64 or u64. Snapshots differ by the type names only.

+++ acvus-ast/src/token.rs
@@ -72,8 +72,8 @@ pub enum Token {
     // -- Literals --
     #[regex(r"[0-9]+\.[0-9]+", |lex| lex.slice().parse::<f64>().ok())]
     FloatLit(f64),
-    #[regex(r"[0-9]+", |lex| lex.slice().parse::<i64>().ok())]
-    IntLit(i64),
+    #[regex(r"[0-9]+", |lex| lex.slice().parse::<i128>().ok())]
+    IntLit(i128),
     #[regex(r#""([^"\\]|\\.)*""#, parse_string_literal)]
     StringLit(String),
